requirements.txt
anthropic>=0.40.0,<1
cryptography>=42.0.0,<43
email-validator>=2.1.0,<3
flask>=3.0.0,<4
flask-bcrypt>=1.0.1,<2
flask-login>=0.6.0,<1
flask-migrate>=4.0.0,<5
flask-sqlalchemy>=3.1.0,<4
flask-wtf>=1.2.0,<2
gunicorn>=22.0.0,<23
psycopg2-binary>=2.9.0,<3
python-dotenv>=1.0.0,<2

